Implemented the documents API and integration tests

+++ BSharp/Entities/Document.cs
@@ -95,7 +95,7 @@ namespace BSharp.Entities
             DocStateName.Reviewed,
             DocStateName.Closed
         })]
-        public int? State { get; set; }
+        public short? State { get; set; }
 
         [Display(Name = "CreatedAt")]
         public DateTimeOffset? CreatedAt { get; set; }
@@ -146,16 +146,16 @@ namespace BSharp.Entities
 
     public static class DocState
     {
-        public const int Draft = 0;
-        public const int Void = -1;
-        public const int Requested = 1;
-        public const int Rejected = -2;
-        public const int Authorized = 2;
-        public const int Failed = -3;
-        public const int Completed = 3;
-        public const int Invalid = -4;
-        public const int Reviewed = 4;
-        public const int Closed = 5;
+        public const short Draft = 0;
+        public const short Void = -1;
+        public const short Requested = 1;
+        public const short Rejected = -2;
+        public const short Authorized = 2;
+        public const short Failed = -3;
+        public const short Completed = 3;
+        public const short Invalid = -4;
+        public const short Reviewed = 4;
+        public const short Closed = 5;
     }
 
     public static class DocStateName
